+++ breptopo_c/HistGraph.h
@@ -19,8 +19,8 @@ class HistGraph
 public:
 	HistGraph();
 
-	uint32_t NextOpId();
-	void Update(const partgraph::BRepHistory& hist, uint32_t op_id);
+	uint16_t NextOpId();
+	void Update(const partgraph::BRepHistory& hist, uint16_t op_id);
 
 	std::shared_ptr<graph::Node> QueryNode(const std::shared_ptr<partgraph::TopoShape>& shape) const;
 	bool QueryNodes(uint32_t uid, std::vector<std::shared_ptr<graph::Node>>& results) const;
@@ -33,7 +33,7 @@ private:
 private:
 	std::shared_ptr<graph::Graph> m_graph;
 
-	uint32_t m_next_op = 0;
+	uint16_t m_next_op = 0;
 
 	// map shape to gid
 	NCollection_DataMap<TopoDS_Shape, size_t, TopTools_ShapeMapHasher> m_curr_shapes;
